newComparer now checks to see if a gmae with that ID already exists in the storage layer.
If it does, then instead of working with a fresh game, as expected, we'll work on top of an existing game, giving weird state.

This is important because RegenerateGolden() needs to only be called once in a given storage layer; this verifies that will be the case.

Part of #701. Part of #755.